+++ ExpandWorldPrefabs/InfoManager.cs
@@ -94,10 +94,10 @@ public class InfoManager
   public static void Patch()
   {
     var canPatch = !Helper.IsClient();
-    var shouldPersistPlayers = canPatch && Settings.PersistPlayers;
-    var shouldRestoreScale = canPatch && Settings.RestoreScale;
-    var shouldSupportAttach = canPatch && Settings.SupportAttach;
-    var shouldServerSideData = canPatch && Settings.ServerSideData;
+    var shouldPersistPlayers = canPatch && Config.PersistPlayers;
+    var shouldRestoreScale = canPatch && Config.RestoreScale;
+    var shouldSupportAttach = canPatch && Config.SupportAttach;
+    var shouldServerSideData = canPatch && Config.ServerSideData;
     PersistPlayers.Patch(EWP.Harmony, shouldPersistPlayers);
     RestoreScale.Patch(EWP.Harmony, shouldRestoreScale);
     SupportAttach.Patch(EWP.Harmony, shouldSupportAttach);