+++ ExpandWorldPrefabs/service/data/ZdoEntry.cs
@@ -13,6 +13,7 @@ namespace Data;
 public class ZdoEntry(int Prefab, Vector3 Position, Vector3 rotation, ZDO zdo)
 {
   // Nulls add more code but should be more performant.
+  public int Prefab = Prefab;
   public Dictionary<int, string>? Strings;
   public Dictionary<int, string>? ServerStrings;
   public Dictionary<int, float>? Floats;
@@ -76,7 +77,7 @@ public class ZdoEntry(int Prefab, Vector3 Position, Vector3 rotation, ZDO zdo)
     zdo.m_prefab = prefab;
     zdo.m_rotation = rotation;
     // Usually players are non persistent but this way NPCs can be spawned without having to manually set persistent in the data.
-    if (prefab == PlayerHash && ExpandWorld.Prefab.Settings.PersistPlayers)
+    if (prefab == PlayerHash && ExpandWorld.Prefab.Config.PersistPlayers)
       zdo.Persistent = true;
     else
       zdo.Persistent = view.m_persistent;